+++ script/init_db/presidents.example.csv
@@ -1,2 +1,3 @@
 email,name
 bot@discord.com,Discord Bot
+deposit.app@scsc.dev,Deposit App
