+++ src/controller/__init__.py
@@ -1,5 +1,5 @@
 from .article import create_article_ctrl, BodyCreateArticle
 from .sig import create_sig_ctrl, BodyCreateSIG, update_sig_ctrl, BodyUpdateSIG
 from .pig import create_pig_ctrl, BodyCreatePIG, update_pig_ctrl, BodyUpdatePIG
-from .user import create_user_ctrl, BodyCreateUser, enroll_user_ctrl, register_oldboy_applicant_ctrl, process_oldboy_applicant_ctrl, reactivate_oldboy_ctrl, verify_enroll_user_ctrl
+from .user import create_user_ctrl, BodyCreateUser, enroll_user_ctrl, register_oldboy_applicant_ctrl, process_oldboy_applicant_ctrl, reactivate_oldboy_ctrl, verify_enroll_user_ctrl, ProcessDepositResult, process_deposit_ctrl
 from .scsc import update_scsc_global_status_ctrl, ctrl_status_available, map_semester_name
