+++ src/model/__init__.py
@@ -4,5 +4,5 @@ from .major import Major
 from .scsc_global_status import SCSCStatus, SCSCGlobalStatus
 from .pig import PIG, PIGMember
 from .sig import SIG, SIGMember
-from .user import User, UserRole, UserStatus, StandbyReqTbl, OldboyApplicant
+from .user import User, UserResponse, UserRole, UserStatus, StandbyReqTbl, OldboyApplicant
 from .comment import Comment
